backend/requirements.txt
Pillow

Django>=5.0
djangorestframework>=3.14,<4.0
dj-rest-auth>=2.1,<3.0
django-allauth==0.50.0
djangorestframework-simplejwt>=5.2,<6.0
social-auth-app-django>=5.0,<6.0
drf-yasg>=1.21,<2.0
psycopg2-binary>=2.9,<3.0
python-decouple>=3.7,<4.0
pytest-asyncio>=0.23.2
pytest-django>=4.7.0
django-cors-headers
google-auth
stripe>=5.0.0